exps_on_image_datasets/requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=4.5=1_gnu
absl-py=0.12.0=pypi_0
aiohttp=3.8.1=pypi_0
aiosignal=1.2.0=pypi_0
alembic=1.5.5=pypi_0
antlr4-python3-runtime=4.8=pypi_0
anykeystore=0.2=pypi_0
apex=0.1=pypi_0
argon2-cffi=20.1.0=pypi_0
ase=3.21.1=pypi_0
async-generator=1.10=pypi_0
async-timeout=4.0.2=pypi_0
asynctest=0.13.0=pypi_0
attrs=20.3.0=pypi_0
autopep8=1.5.7=pyhd3eb1b0_0
backcall=0.2.0=pyhd3eb1b0_0
blas=1.0=mkl
bleach=3.3.0=pypi_0
boto3=1.17.8=pypi_0
botocore=1.20.8=pypi_0
brotlipy=0.7.0=py36h27cfd23_1003
bzip2=1.0.8=h7b6447c_0
ca-certificates=2021.10.26=h06a4308_2
cached-property=1.5.2=pypi_0
cachetools=4.2.1=pypi_0
certifi=2021.5.30=py36h06a4308_0
cffi=1.14.4=pypi_0
charset-normalizer=2.0.4=pyhd3eb1b0_0
click=7.1.2=pypi_0
cliff=3.7.0=pypi_0
cmaes=0.8.2=pypi_0
cmd2=1.5.0=pypi_0
colorama=0.4.4=pypi_0
colorlog=4.7.2=pypi_0
contextlib2=21.6.0=pypi_0
cox=0.1.post3=pypi_0
cryptacular=1.6.2=pypi_0
cryptography=35.0.0=py36hd23ed53_0
cudatoolkit=11.3.1=h2bc3f7f_2
cycler=0.10.0=pypi_0
cython=0.29.21=pypi_0
dataclasses=0.8=pyh4f3eec9_6
datasets=1.17.0=pypi_0
decorator=5.1.0=pyhd3eb1b0_0
defusedxml=0.6.0=pypi_0
dill=0.3.3=pypi_0
entrypoints=0.3=pypi_0
fairseq=0.10.2=pypi_0
ffmpeg=4.3=hf484d3e_0
filelock=3.3.1=pyhd3eb1b0_1
freetype=2.11.0=h70c0345_0
frozenlist=1.2.0=pypi_0
fsspec=2022.1.0=pypi_0
future=0.18.2=pypi_0
gitdb=4.0.7=pypi_0
gitpython=3.1.14=pypi_0
gmp=6.2.1=h2531618_2
gnutls=3.6.15=he1e5248_0
google-auth=1.28.0=pypi_0
google-auth-oauthlib=0.4.4=pypi_0
googledrivedownloader=0.4=pypi_0
gputil=1.4.0=pypi_0
grpcio=1.37.0=pypi_0
h5py=3.1.0=pypi_0
huggingface-hub=0.2.1=pypi_0
huggingface_hub=0.0.17=py_0
hupper=1.10.3=pypi_0
hydra-core=1.0.5=pypi_0
idna=3.3=pyhd3eb1b0_0
idna-ssl=1.1.0=pypi_0
imageio=2.9.0=pypi_0
importlib=1.0.4=pypi_0
importlib-metadata=3.4.0=pypi_0
importlib-resources=5.1.0=pypi_0
importlib_metadata=4.8.1=hd3eb1b0_0
install=1.3.4=pypi_0
intel-openmp=2021.4.0=h06a4308_3561
ipykernel=5.3.4=py36h5ca1d4c_0
ipython=7.16.1=py36h5ca1d4c_0
ipython_genutils=0.2.0=pyhd3eb1b0_1
ipywidgets=7.6.3=pypi_0
isodate=0.6.0=pypi_0
jedi=0.17.0=py36_0
jinja2=2.11.3=pypi_0
joblib=1.0.0=pypi_0
jpeg=9b=h024ee3a_2
jsonschema=3.0.2=pypi_0
jupyter_client=7.0.6=pyhd3eb1b0_0
jupyter_core=4.8.1=py36h06a4308_0
jupyterlab-pygments=0.1.2=pypi_0
jupyterlab-widgets=1.0.0=pypi_0
kiwisolver=1.3.1=pypi_0
lame=3.100=h7b6447c_0
lcms2=2.12=h3be6417_0
ld_impl_linux-64=2.35.1=h7274673_9
libffi=3.3=he6710b0_2
libgcc-ng=9.3.0=h5101ec6_17
libgomp=9.3.0=h5101ec6_17
libiconv=1.15=h63c8f33_5
libidn2=2.3.2=h7f8727e_0
libpng=1.6.37=hbc83047_0
libprotobuf=3.17.2=h4ff587b_1
libsodium=1.0.18=h7b6447c_0
libstdcxx-ng=9.3.0=hd4cf53a_17
libtasn1=4.16.0=h27cfd23_0
libtiff=4.2.0=h85742a9_0
libunistring=0.9.10=h27cfd23_0
libuv=1.40.0=h7b6447c_0
libwebp-base=1.2.0=h27cfd23_0
llvmlite=0.35.0=pypi_0
lz4-c=1.9.3=h295c915_1
mako=1.1.4=pypi_0
markdown=3.3.4=pypi_0
markupsafe=1.1.1=pypi_0
matplotlib=3.3.4=pypi_0
mistune=0.8.4=pypi_0
mkl=2020.2=256
mkl-service=2.3.0=py36he8ac12f_0
mkl_fft=1.2.0=py36h23d657b_0
mkl_random=1.1.1=py36h0573a6f_0
ml-collections=0.1.0=pypi_0
multidict=5.2.0=pypi_0
multiprocess=0.70.11.1=pypi_0
nbclient=0.5.2=pypi_0
nbconvert=6.0.7=pypi_0
nbformat=5.1.2=pypi_0
ncurses=6.3=h7f8727e_2
nest-asyncio=1.5.1=pyhd3eb1b0_0
nettle=3.7.3=hbbd107a_1
networkx=2.5=pypi_0
ninja=1.10.2=h5e70eb0_2
nlpaug=1.1.2=pypi_0
notebook=6.2.0=pypi_0
numba=0.52.0=pypi_0
numexpr=2.7.3=pypi_0
numpy=1.19.2=py36h54aff64_0
numpy-base=1.19.2=py36hfa32c7d_0
oauthlib=3.1.0=pypi_0
olefile=0.46=py36_0
omegaconf=2.0.6=pypi_0
openh264=2.1.0=hd408876_0
openssl=1.1.1l=h7f8727e_0
optuna=2.5.0=pypi_0
packaging=20.8=pypi_0
pandas=1.1.5=pypi_0
pandocfilters=1.4.3=pypi_0
parso=0.8.2=pyhd3eb1b0_0
pastedeploy=2.1.1=pypi_0
pbkdf2=1.3=pypi_0
pbr=5.5.1=pypi_0
pexpect=4.8.0=pyhd3eb1b0_3
pickleshare=0.7.5=pyhd3eb1b0_1003
pillow=8.1.0=py36he98fc37_0
pip=21.2.2=py36h06a4308_0
plaster=1.0=pypi_0
plaster-pastedeploy=0.7=pypi_0
portalocker=2.1.0=pypi_0
pqdict=1.1.1=pypi_0
prettytable=2.1.0=pypi_0
prometheus-client=0.9.0=pypi_0
prompt-toolkit=3.0.20=pyhd3eb1b0_0
protobuf=3.17.2=py36h295c915_0
psutil=5.8.0=pypi_0
ptyprocess=0.7.0=pyhd3eb1b0_2
py3nvml=0.2.6=pypi_0
pyarrow=3.0.0=pypi_0
pyasn1=0.4.8=pypi_0
pyasn1-modules=0.2.8=pypi_0
pycodestyle=2.8.0=pyhd3eb1b0_0
pycparser=2.21=pyhd3eb1b0_0
pygments=2.10.0=pyhd3eb1b0_0
pyopenssl=21.0.0=pyhd3eb1b0_1
pyparsing=3.0.4=pyhd3eb1b0_0
pyperclip=1.8.2=pypi_0
pyramid=2.0=pypi_0
pyramid-mailer=0.15.1=pypi_0
pyrsistent=0.17.3=pypi_0
pysocks=1.7.1=py36h06a4308_0
python=3.6.12=hcff3b4d_2
python-dateutil=2.8.2=pyhd3eb1b0_0
python-editor=1.0.4=pypi_0
python-louvain=0.15=pypi_0
python3-openid=3.2.0=pypi_0
python_abi=3.6=1_cp36m
pytorch=1.10.0=py3.6_cuda11.3_cudnn8.2.0_0
pytorch-lightning=1.0.5=pypi_0
pytorch-mutex=1.0=cuda
pytorch-pretrained-bert=0.6.2=pypi_0
pytz=2021.1=pypi_0
pywavelets=1.1.1=pypi_0
pyyaml=5.4.1=pypi_0
pyzmq=22.2.1=py36h295c915_1
rdflib=5.0.0=pypi_0
readline=8.1=h27cfd23_0
regex=2021.8.3=py36h7f8727e_0
repoze-sendmail=4.4.1=pypi_0
requests=2.26.0=pyhd3eb1b0_0
requests-oauthlib=1.3.0=pypi_0
robustness=1.2.1.post2=pypi_0
rsa=4.7.2=pypi_0
s3transfer=0.3.4=pypi_0
sacrebleu=1.5.0=pypi_0
sacremoses=0.0.43=pypi_0
scikit-image=0.17.2=pypi_0
scikit-learn=0.24.1=pypi_0
scipy=1.5.4=pypi_0
seaborn=0.11.1=pypi_0
send2trash=1.5.0=pypi_0
setuptools=58.0.4=py36h06a4308_0
six=1.16.0=pyhd3eb1b0_0
smmap=4.0.0=pypi_0
sqlalchemy=1.3.23=pypi_0
sqlite=3.36.0=hc218d9a_0
stevedore=3.3.0=pypi_0
tables=3.6.1=pypi_0
tensorboard=2.4.1=pypi_0
tensorboard-logger=0.1.0=pypi_0
tensorboard-plugin-wit=1.8.0=pypi_0
tensorboardx=2.2=pypi_0
terminado=0.9.2=pypi_0
testpath=0.4.4=pypi_0
threadpoolctl=2.1.0=pypi_0
tifffile=2020.9.3=pypi_0
tk=8.6.11=h1ccaba5_0
tokenizers=0.9.4=pypi_0
toml=0.10.2=pyhd3eb1b0_0
torch-cluster=1.5.9=pypi_0
torch-geometric=1.6.3=pypi_0
torch-sparse=0.6.9=pypi_0
torch-spline-conv=1.2.1=pypi_0
torchaudio=0.10.0=py36_cu113
torchvision=0.11.1=py36_cu113
tornado=6.1=py36h27cfd23_0
tqdm=4.49.0=pypi_0
traitlets=4.3.3=py36h06a4308_0
transaction=3.0.1=pypi_0
transformers=4.3.2=pypi_0
translationstring=1.4=pypi_0
typer=0.3.2=pypi_0
typing-extensions=3.10.0.2=hd3eb1b0_0
typing_extensions=3.10.0.2=pyh06a4308_0
urllib3=1.26.7=pyhd3eb1b0_0
velruse=1.1.1=pypi_0
venusian=3.0.0=pypi_0
wcwidth=0.2.5=pyhd3eb1b0_0
webencodings=0.5.1=pypi_0
webob=1.8.7=pypi_0
werkzeug=1.0.1=pypi_0
wheel=0.37.0=pyhd3eb1b0_1
widgetsnbextension=3.5.1=pypi_0
wtforms=2.3.3=pypi_0
wtforms-recaptcha=0.3.2=pypi_0
xmltodict=0.12.0=pypi_0
xxhash=2.0.0=pypi_0
xz=5.2.5=h7b6447c_0
yarl=1.7.2=pypi_0
zeromq=4.3.4=h2531618_0
zipp=3.4.0=pypi_0
zlib=1.2.11=h7b6447c_3
zope-deprecation=4.4.0=pypi_0
zope-interface=5.4.0=pypi_0
zope-sqlalchemy=1.5=pypi_0
zstd=1.4.9=haebb681_0
